feat: Transfer models and repositories to new project
Initialized a new project and successfully transferred existing models and repositories from a previous project. This setup lays the groundwork for integrating a robust logging framework to enhance the application's monitoring and debugging capabilities. Preparing to establish a comprehensive logging infrastructure next.

+++ Tasks/Task3.3/ProductLogging.Application/Services/ProductService.cs
@@ -1,26 +1,29 @@
 ﻿using AutoMapper;
+using ProductLogging.Application.Contracts;
 using ProductLogging.Application.Interfaces;
 using ProductLogging.Dtos;
 using ProductLogging.Infrastracture.Interface;
 using ProductLogging.Models;
 
 namespace ProductLogging.Application.Services;
-
 public class ProductService : IProductService
 {
     private readonly IProductRepository _productRepository;
     private readonly IMapper _mapper;
+    private readonly ILoggerService _loggerService;
 
-    public ProductService(IProductRepository productRepository, IMapper mapper)
+    public ProductService(IProductRepository productRepository, IMapper mapper, ILoggerService loggerService)
     {
         _productRepository = productRepository;
         _mapper = mapper;
+        _loggerService = loggerService;
     }
 
-    public async Task AddAsync(ProductDto productDto)
+    public async Task AddAsync(ProductCreateDto productDto)
     {
         var product = _mapper.Map<Product>(productDto);
         await _productRepository.AddAsync(product);
+      
     }
 
     public async Task<IEnumerable<ProductDto>> GetAllAsync()
@@ -38,12 +41,27 @@ public class ProductService : IProductService
     public async Task UpdateAsync(ProductDto productDto)
     {
         var product = await _productRepository.GetByIdAsync(productDto.Id);
+        if (product is null)
+        {
+            string message = $"The Book with id :{productDto.Id} could not found";
+            _loggerService.LogInfo(message);
+            throw new Exception(message);
+        }
         _mapper.Map(productDto, product);
         await _productRepository.UpdateAsync(product);
     }
 
+
     public async Task DeleteAsync(int id)
     {
+        var product = await _productRepository.GetByIdAsync(id);
+        if(product is null)
+        {
+            string message = $"The Book with id :{id} could not found";
+            _loggerService.LogInfo(message);
+            throw new Exception(message);
+        }
         await _productRepository.DeleteAsync(id);
+
     }
 }
